implement new save data system
For the mirror world reverse music, and maybe a few other items, I think we need to use save data variables. This commit changes the current system of saving the "usernolayers" variable as a bit to saving in a lookup table encoded with JSON. I tested that legacy save data files (with a 0 or 1) are successfully converted.