add basic functionality for Knife Task music
Using the special TargetRoomIndexes, play the reversed Downpour and Dross themes when in the mirrored world, and the mineshaft ambient and mineshaft escape music when in the mineshaft. The former works well, while the latter needs a few touch-ups (the correct music doesn't play when Mother's Shadow spawns, also the ambient track restarted upon entering the Knife Piece 2 room).

We leverage save data so that the correct music will play even upon closing the game and then continuing the run.